requirements.txt
geopandas
matplotlib
requests
Pillow
pycountry